+++ crates/trie/db/src/lib.rs
@@ -1,13 +1,21 @@
 //! An integration of [`reth-trie`] with [`reth-db`].
 
+mod hashed_cursor;
 mod prefix_set;
 mod proof;
 mod state;
 mod storage;
+mod trie_cursor;
 mod witness;
 
+pub use hashed_cursor::{
+    DatabaseHashedAccountCursor, DatabaseHashedCursorFactory, DatabaseHashedStorageCursor,
+};
 pub use prefix_set::PrefixSetLoader;
 pub use proof::DatabaseProof;
 pub use state::{DatabaseHashedPostState, DatabaseStateRoot};
 pub use storage::DatabaseStorageRoot;
+pub use trie_cursor::{
+    DatabaseAccountTrieCursor, DatabaseStorageTrieCursor, DatabaseTrieCursorFactory,
+};
 pub use witness::DatabaseTrieWitness;
